fix app/feat/repository

See merge request s09-final/S09P31A207!77

+++ frontend/gunpang/data/src/main/java/com/gunpang/data/model/dto/AvatarDataModel.kt
@@ -15,13 +15,15 @@ data class AvatarDetailDto(
     val deathCause: DeathCauseCode, // 사망 원인
 )
 
+
 data class CurrentAvatarInfoDto(
     val avatarId: Int,
     val avatarType: Int,
     val name: String, // 이름
     val stage: String, // 배경 단계
     val status: AvatarStatusCode, // 상태(생존/사망/졸업)
-    val healthPoint: Int, // 체력
+    val healthPoint: Float, // 체력
     val startedDate: String, // 시작 날짜
     val finishedDate: String, // 종료 날짜
 )
+
